requirements.txt
binance-connector==3.5.1
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
numpy==2.1.2
click==8.1.7
PyYAML==6.0.2
